add: ResultTitle component
